change time scale from int to long

+++ chemclipse/plugins/org.eclipse.chemclipse.nmr.model/src/org/eclipse/chemclipse/nmr/model/support/ISignalExtractor.java
@@ -37,9 +37,9 @@ public interface ISignalExtractor {
 
 	double[] extractIntesityFIDReal();
 
-	int[] extractTimeFID();
+	long[] extractTimeFID();
 
-	void createScansFID(Complex[] complexFID, int[] time);
+	void createScansFID(Complex[] complexFID, long[] time);
 
 	void createScans(Complex[] fourierTransformedData, double[] chemicalShift);
